backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.7.1
async-timeout==4.0.3
attrs==25.3.0
Authlib==1.3.1
beautifulsoup4==4.13.4
bidict==0.23.1
blessed==1.21.0
blis==0.7.10
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.21.1
cohere==5.15.0
confection==0.1.5
cryptography==45.0.3
cymem==2.0.11
dataclasses-json==0.6.7
deep-translator==1.11.4
deepl==1.22.0
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
editor==1.6.6
exceptiongroup==1.3.0
fastapi==0.115.12
fastavro==1.11.1
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.5.1
grpcio==1.73.0
grpcio-health-checking==1.73.0
grpcio-tools==1.73.0
h11==0.16.0
hf-xet==1.1.3
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.4
idna==3.10
iniconfig==2.1.0
inquirer==3.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.1.17
langchain-community==0.0.37
langchain-core==0.1.51
langchain-text-splitters==0.0.2
langcodes==3.5.0
langsmith==0.1.147
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.4
murmurhash==1.0.13
mypy_extensions==1.1.0
networkx==3.2.1
numexpr==2.10.2
numpy==1.26.4
openai==0.28
orjson==3.10.18
packaging==23.2
pillow==11.2.1
pluggy==1.6.0
preshed==3.0.10
propcache==0.3.2
protobuf==6.31.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.1
pymongo==4.13.0
pytest==8.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.2
python-socketio==5.13.0
PyYAML==6.0.2
readchar==4.2.1
redis==4.6.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rq==1.15.1
runs==1.2.2
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==4.1.0
shellingham==1.5.4
simple-websocket==1.1.0
simsimd==6.4.9
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.4.54
srsly==2.5.1
starlette==0.46.2
sympy==1.14.0
tabulate==0.9.0
tenacity==8.5.0
thinc==8.2.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.7.1
tqdm==4.67.1
transformers==4.52.4
typer==0.16.0
types-PyYAML==6.0.12.20250516
types-requests==2.32.4.20250611
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
validators==0.34.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
weaviate-client==3.26.7
wrapt==1.17.2
wsproto==1.2.0
xmod==1.8.1
yarl==1.20.1
zstandard==0.23.0
kiwipiepy==0.20.3